requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
pydantic==2.4.2
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.0
prophet==1.1.4
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0